requirement.txt
absl-py==0.7.1
astor==0.7.1
atari-py==0.1.7
atomicwrites==1.2.1
attrs==18.2.0
backcall==0.1.0
-e git+https://github.com/tunglm2203/rl_collections.git@f815a0736a1f91a1162a21ae6abf3707d7983399#egg=baselines&subdirectory=external_packages/baselines
bleach==1.5.0
box2d-py==2.3.8
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudpickle==0.5.2
colorama==0.4.1
cycler==0.10.0
Cython==0.29.3
decorator==4.4.0
dill==0.2.9
-e git+https://github.com/WarrG3X/dobot-rl.git@21a011679c64a8ef79c82df755146f33d6bf8a56#egg=dobot_rl
enum34==1.1.6
filelock==3.0.10
funcsigs==1.0.2
future==0.17.1
gast==0.2.2
glfw==1.7.0
glob2==0.6
grpcio==1.20.1
-e git+https://github.com/openai/gym.git@12e8b763d5dcda4962cbd17887d545f0eec6808a#egg=gym
-e git+https://github.com/WarrG3X/gym-dobot@5ef285e0c3a5786b9783b7230c09e8e8fe1e29a6#egg=gym_dobot
h5py==2.9.0
html5lib==0.9999999
idna==2.8
imageio==2.4.1
ipython==7.4.0
ipython-genutils==0.2.0
jedi==0.13.3
joblib==0.13.1
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
lockfile==0.12.2
Markdown==3.1
matplotlib==3.0.2
mock==2.0.0
more-itertools==5.0.0
mpi4py==3.0.0
mujoco-py==1.50.1.68
numpy==1.16.0
opencv-python==4.0.0.21
pandas==0.24.2
parso==0.4.0
pathlib2==2.3.3
pbr==5.1.3
pexpect==4.7.0
pickleshare==0.7.5
Pillow==5.3.0
pluggy==0.8.1
progressbar2==3.39.2
prompt-toolkit==2.0.9
protobuf==3.9.1
psutil==5.6.1
pthflops==0.1.0
ptyprocess==0.6.0
py==1.7.0
pycparser==2.19
pyglet==1.3.2
Pygments==2.3.1
PyOpenGL==3.1.0
pyparsing==2.3.0
pytest==4.2.0
python-dateutil==2.7.5
python-utils==2.3.0
pytz==2019.1
PyYAML==5.1.2
pyzmq==17.1.2
ray==0.7.3
redis==3.3.8
requests==2.21.0
scipy==1.2.0
seaborn==0.8.1
setproctitle==1.1.10
six==1.12.0
-e git+https://github.com/openai/spinningup.git@97c8c342c45e5bb51005a8515df23ba9c48f0782#egg=spinup
tensorboard==1.12.2
tensorboardX==1.6
tensorflow==1.12.0
tensorflow-gpu==1.12.0
termcolor==1.1.0
thop==0.0.22
torch==0.4.0
torchvision==0.2.1
tqdm==4.28.1
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
Werkzeug==0.15.2
zmq==0.0.0
